+++ pkg/dataq_test.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"reflect"
@@ -7,65 +7,6 @@ import (
 	"github.com/Knetic/govaluate"
 	log "github.com/sirupsen/logrus"
 )
-
-type Level2 struct {
-	Ypsilon int
-	Omega   string
-}
-
-type Level1 struct {
-	Alfa  float64
-	beta  string
-	Gamma *Level2
-	Zeta  map[string]float64
-}
-
-const (
-	SEP                   = "."
-	Alfa_name             = "Alfa"
-	Alfa_value            = 1.0
-	Alfa_update           = 2.0
-	Gamma                 = "Gamma"
-	Ypsilon_name          = "Ypsilon"
-	Ypsilon_value         = 10
-	Omega_name            = "Omega"
-	Omega_value           = "test2"
-	Omega_update          = "updated"
-	Beta_name             = "beta"
-	Beta_value            = "test1"
-	Expression            = "Alfa + Gamma_Ypsilon"
-	Expression_result     = 11.0
-	Zeta                  = "Zeta"
-	Zeta_field1           = "zeta1"
-	Zeta_field1_value     = 1.0
-	Zeta_field2           = "zeta2"
-	Zeta_field2_value     = 2.0
-	Zeta_supported_fields = 2
-)
-
-var Vars = map[string]interface{}{
-	Alfa_name:                  Alfa_value,
-	Gamma + SEP + Ypsilon_name: Ypsilon_value,
-	Gamma + SEP + Omega_name:   Omega_value,
-	Zeta + SEP + Zeta_field1:   Zeta_field1_value,
-	Zeta + SEP + Zeta_field2:   Zeta_field2_value,
-}
-
-func getData() Level1 {
-	l2 := Level2{
-		Ypsilon: Ypsilon_value,
-		Omega:   Omega_value,
-	}
-	return Level1{
-		Alfa:  Alfa_value,
-		beta:  Beta_value,
-		Gamma: &l2,
-		Zeta: map[string]float64{
-			Zeta_field1: Zeta_field1_value,
-			Zeta_field2: Zeta_field2_value,
-		},
-	}
-}
 
 func TestGetVars(t *testing.T) {
 	l1 := getData()
